# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta es la definición de la API del servidor del Hydro Monitor",
        "title": "Hydro Monitor Web API",
        "contact": {
            "name": "Manuel Porto",
            "email": "manu.porto94@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.0"
    },
    "host": "localhost",
    "basePath": "/api",
    "paths": {
...
        "/nodes/last-reading": {
            "get": {
                "tags": [
                    "readings"
                ],
                "summary": "Obtiene la última medición de todos los nodos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/api_models.GetReadingDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/nodes/{node_id}/readings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "readings"
                ],
                "summary": "Obtiene las mediciones de un nodo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Cantidad máxima de mediciones por página",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "String en base 64 que contiene el estado de pagina. Utilizado para traer la próxima página",
                        "name": "page_state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api_models.GetReadingDTO"
                            }
                        },
                        "headers": {
                            "Page-State": {
                                "type": "string",
                                "description": "Page state"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "readings"
                ],
                "summary": "Crea una medición",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos de la medición",
                        "name": "reading",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api_models.ReadingDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_models.GetReadingDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/nodes/{node_id}/readings/{reading_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "readings"
                ],
                "summary": "Obtiene los datos de una medición",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de la medición",
                        "name": "reading_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_models.GetReadingDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "readings"
                ],
                "summary": "Borra una medición",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de la medición",
                        "name": "reading_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/nodes/{node_id}/readings/{reading_id}/photos": {
            "get": {
                "produces": [
                    "image/jpeg"
                ],
                "tags": [
                    "readings"
                ],
                "summary": "Obtiene la foto de una medición",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de la medición",
                        "name": "reading_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "image/jpeg"
                ],
                "tags": [
                    "readings"
                ],
                "summary": "Agrega una foto a la medición",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del nodo",
                        "name": "node_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de la medición",
                        "name": "reading_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Foto de la medición",
                        "name": "picture",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api_models.GetReadingDTO": {
            "type": "object",
            "properties": {
                "manualReading": {
                    "type": "boolean",
                    "example": false
                },
                "readingId": {
                    "type": "string",
                    "example": "00336270-8191-11ea-a43d-0242ac120003"
                },
                "readingTime": {
                    "type": "string",
                    "example": "2020-04-26T19:47:53.391Z"
                },
                "waterLevel": {
                    "type": "number",
                    "example": 60
                }
            }
        },
...
        "api_models.ReadingDTO": {
            "type": "object",
            "properties": {
                "manualReading": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-04-26T19:47:53.391Z"
                },
                "waterLevel": {
                    "type": "number",
                    "example": 60
                }
            }
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "internal": {
                    "description": "Stores the error returned by an external dependency",
                    "type": "error"
                },
                "message": {
                    "type": "object"
                }
            }
        }
    }
}